Refactor: (Feedback) LadderGameController 생성 (기존의 Main 클래스 대체), LadderGame 정적 팩토리 메서드 추가
