+++ src/main/java/nextstep/ladder/controller/LadderGameController.java
@@ -1,23 +1,19 @@
-package nextstep.ladder;
+package nextstep.ladder.controller;
 
-import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.LadderGame;
-import nextstep.ladder.domain.Participants;
 import nextstep.ladder.ui.InputView;
 import nextstep.ladder.ui.ResultView;
 
 import java.util.List;
 
-public class Main {
+public class LadderGameController {
 
     public static void main(String[] args) {
         try {
             List<String> names = InputView.readParticipantNames();
             int height = InputView.readLadderMaxHeight();
 
-            Participants participants = Participants.of(names);
-            Ladder ladder = Ladder.of(participants.numberOf(), height);
-            LadderGame ladderGame = new LadderGame(participants, ladder);
+            LadderGame ladderGame = LadderGame.of(names, height);
 
             ResultView.printGameResult(ladderGame);
         } catch (IllegalArgumentException e) {
